revising 2 to help 3

+++ 0x1A-hash_tables/2-key_index.c
@@ -9,7 +9,7 @@
  */
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	int index;
+	unsigned long int index;
 
 	index = hash_djb2(key);
 	index = index % size;
